TSDK-826 Refactor
